Moved pkg/config to internal directory
